requirements.txt
Faker>=20.0.0
PyYAML>=6.0
click>=8.0
Django>=3.2
pytest
